Rename most commands to share a name: mark
j is and important time saver, so I'm keeping it as j. The other
commands are infrequently used, so using single characters makes it too
hard to remember. The system is called bashmarks, so use mark as the
common name.

Renamed all commands, updated the documentation, and also renamed some
internal functions to have more descriptive names.

`mark` sets the current dir to be marked by the input name and on its
own it lists the current marks. It also accepts the help commands.

mark_print and mark_delete aren't often used by me, so I'm fine with
long names. Ideally, they could be commands to `mark`?

Tested all commands and tab completion on j, print, and delete.